+++ src/data/inputNames.js
@@ -1,18 +1,16 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
-import { TEXT } from "../common/index";
+import { TEXT } from "../utils/common";
 
 const inputNames = async () => {
   const names = await MissionUtils.Console.readLineAsync(`${TEXT.INPUT_NAMES}`);
-  const NamesIsValid = isValidNames(names);
-  if (!NamesIsValid) throw Error(TEXT.INPUT_ERROR);
+  validateNames(names);
   return names;
 };
 
-const isValidNames = (names) => {
+const validateNames = (names) => {
   for (let name of names.split(",")) {
-    if (name.length > 5) false;
+    if (name.length > 5) throw Error(TEXT.INPUT_ERROR);
   }
-  return true;
 };
 
 export default inputNames;
